+++ dags/scd_stations.py
@@ -7,41 +7,11 @@ from airflow import DAG
 from google.cloud import bigquery
 from airflow.decorators import task
 from airflow.providers.google.cloud.hooks.bigquery import BigQueryHook
-
+from utils.utils import create_schema_fields
 
 table_id = f"{os.environ['GCP_PROJECT_ID']}.{os.environ['GCP_DATASET']}.{os.environ['GCP_TABLE_STATIONS']}"
 bq_hook = BigQueryHook(gcp_conn_id='google_cloud_default', use_legacy_sql=False)
 client_bq = bq_hook.get_client()
-
-def create_schema_fields(schema):
-    def create_field(field):
-        # Default mode and description if not provided
-        mode = field.get('mode', 'NULLABLE')
-        description = field.get('description', None)
-
-        if field['type'] == 'RECORD':
-            # Recursively process sub-fields
-            sub_fields = [
-                create_field(sub_field)
-                for sub_field in field.get('fields', [])
-            ]
-            return bigquery.SchemaField(
-                field['name'],
-                'RECORD',
-                mode=mode,
-                description=description,
-                fields=sub_fields
-            )
-        else:
-            return bigquery.SchemaField(
-                field['name'],
-                field['type'],
-                mode=mode,
-                description=description
-            )
-
-    return [create_field(field) for field in schema]
-
 
 @task
 def extract():
